requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
lxml>=5.0.0
tenacity>=8.2.0
